The type defintion for Acceptor and Value is wrong in the previous commit. They were hardwired to String, now we changed it back.
The reason is the following means Acceptor is a type that contains other smaller types (such as Nat). This is not what we mean by having acceptor as a type variable.

abbrev Acceptor := Type

If we want to make it flexible (not hardwired to String), we have to declare them as variables (which I tried to avoid because they have to be supplied bascially everywhere and that makes the code somewhat messy). Instead, we use String for now.